refactor(dashboard): moving data streaming
