+++ app/ui/dashboard/cards.tsx
@@ -3,8 +3,10 @@ import {
   ClockIcon,
   UserGroupIcon,
   InboxIcon,
-} from '@heroicons/react/24/outline';
-import { lusitana } from '@/app/ui/fonts';
+} from "@heroicons/react/24/outline";
+import prisma from "@/app/lib/prisma";
+import { lusitana } from "@/app/ui/fonts";
+import { formatCurrency } from "@/app/lib/utils";
 
 const iconMap = {
   collected: BanknotesIcon,
@@ -14,18 +16,38 @@ const iconMap = {
 };
 
 export default async function CardWrapper() {
+  const [invoiceCount, customerCount, invoiceStatus] = await Promise.all([
+    prisma.invoice.count(),
+    prisma.customer.count(),
+    prisma.invoice.aggregate({
+      _sum: {
+        amount: true,
+      },
+      _count: {
+        id: true,
+      },
+      where: {
+        OR: [{ status: "paid" }, { status: "pending" }],
+      },
+    }),
+  ]);
+
+  const numberOfInvoices = invoiceCount;
+  const numberOfCustomers = customerCount;
+  const totalPaidInvoices = formatCurrency(invoiceStatus._sum.amount || 0);
+  const totalPendingInvoices = formatCurrency(invoiceStatus._sum.amount || 0);
   return (
     <>
       {/* NOTE: Uncomment this code in Chapter 9 */}
 
-      {/* <Card title="Collected" value={totalPaidInvoices} type="collected" />
+      <Card title="Collected" value={totalPaidInvoices} type="collected" />
       <Card title="Pending" value={totalPendingInvoices} type="pending" />
       <Card title="Total Invoices" value={numberOfInvoices} type="invoices" />
       <Card
         title="Total Customers"
         value={numberOfCustomers}
         type="customers"
-      /> */}
+      />
     </>
   );
 }
@@ -37,7 +59,7 @@ export function Card({
 }: {
   title: string;
   value: number | string;
-  type: 'invoices' | 'customers' | 'pending' | 'collected';
+  type: "invoices" | "customers" | "pending" | "collected";
 }) {
   const Icon = iconMap[type];
