+++ app/ui/dashboard/latest-invoices.tsx
@@ -1,13 +1,33 @@
 import { ArrowPathIcon } from "@heroicons/react/24/outline";
 import clsx from "clsx";
 import Image from "next/image";
+import prisma from "@/app/lib/prisma";
 import { lusitana } from "@/app/ui/fonts";
-import { LatestInvoice } from "@/app/lib/definitions";
-export default async function LatestInvoices({
-  latestInvoices,
-}: {
-  latestInvoices: LatestInvoice[];
-}) {
+import { formatCurrency } from "@/app/lib/utils";
+
+export default async function LatestInvoices() {
+  const invoices = await prisma.invoice.findMany({
+    select: {
+      id: true,
+      amount: true,
+      date: true,
+      customer: {
+        select: {
+          name: true,
+          email: true,
+          imageUrl: true,
+        },
+      },
+    },
+    orderBy: {
+      date: "desc",
+    },
+    take: 5,
+  });
+  const latestInvoices = invoices.map((invoice) => ({
+    ...invoice,
+    amount: formatCurrency(invoice.amount),
+  }));
   return (
     <div className="flex w-full flex-col md:col-span-4">
       <h2 className={`${lusitana.className} mb-4 text-xl md:text-2xl`}>
