+++ app/ui/dashboard/revenue-chart.tsx
@@ -1,7 +1,8 @@
-import { generateYAxis } from "@/app/lib/utils";
-import { CalendarIcon } from "@heroicons/react/24/outline";
+import prisma from "@/app/lib/prisma";
 import { lusitana } from "@/app/ui/fonts";
+import { generateYAxis } from "@/app/lib/utils";
 import { Revenue } from "@/app/lib/definitions";
+import { CalendarIcon } from "@heroicons/react/24/outline";
 
 // This component is representational only.
 // For data visualization UI, check out:
@@ -9,11 +10,8 @@ import { Revenue } from "@/app/lib/definitions";
 // https://www.chartjs.org/
 // https://airbnb.io/visx/
 
-export default async function RevenueChart({
-  revenue,
-}: {
-  revenue: Revenue[];
-}) {
+export default async function RevenueChart() {
+  const revenue = await prisma.revenue.findMany();
   const chartHeight = 350;
   // NOTE: Uncomment this code in Chapter 7
